refactor: use usememo in the right way on the duration card
